+++ .devcontainer/dev/setup-remote-vm.sh
@@ -9,7 +9,7 @@ set -eo pipefail
 BUILDARCH=$([ "$(uname -i)" = "aarch64" ] && echo "arm64" || echo "amd64")
 HOSTNAME="knfsd-dev-ec2"
 USERNAME="ubuntu"
-VERSION="1.1.0-alpha.17"
+VERSION="1.1.0-alpha.18"
 
 ## set env vars for build env only
 export HOME=/home/${USERNAME}
@@ -140,9 +140,9 @@ curl -fsSL "https://awscli.amazonaws.com/awscli-exe-linux-$(uname -i).zip" -o /t
 # https://github.com/bats-core/bats-core/releases
 KNFSD_BATS_CORE_VERSION=1.13.0
 # https://github.com/psf/black/releases
-KNFSD_BLACK_VERSION=25.11.0
+KNFSD_BLACK_VERSION=25.12.0
 # https://github.com/boto/boto3/tags
-KNFSD_BOTO3_VERSION=1.42.0
+KNFSD_BOTO3_VERSION=1.42.6
 # https://hub.docker.com/r/bridgecrew/checkov/tags
 KNFSD_CHECKOV_VERSION=3.2.495
 # https://github.com/codespell-project/codespell/releases
@@ -150,9 +150,9 @@ KNFSD_CODESPELL_VERSION=2.4.1
 # https://github.com/editorconfig-checker/editorconfig-checker/releases
 KNFSD_EDITORCONFIG_VERSION=3.6.0
 # https://github.com/golangci/golangci-lint/releases
-KNFSD_GOLANGCI_LINT_VERSION=2.6.2
+KNFSD_GOLANGCI_LINT_VERSION=2.7.2
 # https://go.dev/dl/
-KNFSD_GOLANG_VERSION=1.25.4
+KNFSD_GOLANG_VERSION=1.25.5
 # https://github.com/python/mypy/tags
 KNFSD_MYPY_VERSION=1.19.0
 # https://github.com/hashicorp/packer/releases
@@ -160,11 +160,11 @@ KNFSD_PACKER_VERSION=1.14.3
 # https://github.com/pre-commit/pre-commit/releases
 KNFSD_PRECOMMIT_VERSION=4.5.0
 # https://pypi.org/project/psycopg/
-KNFSD_PSYCOPG_VERSION=3.3.0
+KNFSD_PSYCOPG_VERSION=3.3.2
 # https://github.com/pylint-dev/pylint/tags
 KNFSD_PYLINT_VERSION=4.0.4
 # https://github.com/semgrep/semgrep/releases
-KNFSD_SEMGREP_VERSION=1.144.0
+KNFSD_SEMGREP_VERSION=1.145.0
 # https://pypi.org/project/shellcheck-py/
 KNFSD_SHELLCHECK_PY_VERSION=0.11.0.1
 # https://github.com/mvdan/sh/releases
@@ -172,13 +172,13 @@ KNFSD_SHFMT_VERSION=3.12.0
 # https://github.com/hashicorp/terraform/releases
 KNFSD_TERRAFORM_VERSION=1.2.9
 # https://github.com/gruntwork-io/terragrunt/releases
-KNFSD_TERRAGRUNT_VERSION=0.93.12
+KNFSD_TERRAGRUNT_VERSION=0.94.0
 # https://github.com/terraform-linters/tflint/releases
 KNFSD_TFLINT_VERSION=0.60.0
 # https://github.com/aquasecurity/tfsec/releases
 KNFSD_TFSEC_VERSION=1.28.14
 # https://github.com/aquasecurity/trivy/releases
-KNFSD_TRIVY_VERSION=0.67.2
+KNFSD_TRIVY_VERSION=0.68.1
 
 ## install golang, delete empty lines and lines containing PATH= in /etc/environment
 curl -fsSL "https://dl.google.com/go/go${KNFSD_GOLANG_VERSION}.linux-${BUILDARCH}.tar.gz" -o "/tmp/go${KNFSD_GOLANG_VERSION}.linux-${BUILDARCH}.tar.gz" \